+++ cli/client/testing/mock_silenced_client.go
@@ -15,19 +15,19 @@ func (c *MockClient) UpdateSilenced(silenced *types.Silenced) error {
 }
 
 // DeleteSilenced for use with mock lib
-func (c *MockClient) DeleteSilenced(subscription, check string) error {
-	args := c.Called(subscription, check)
+func (c *MockClient) DeleteSilenced(id string) error {
+	args := c.Called(id)
 	return args.Error(0)
 }
 
 // FetchSilenced for use with mock lib
-func (c *MockClient) FetchSilenced(subscription, check string) (*types.Silenced, error) {
-	args := c.Called(subscription, check)
+func (c *MockClient) FetchSilenced(id string) (*types.Silenced, error) {
+	args := c.Called(id)
 	return args.Get(0).(*types.Silenced), args.Error(1)
 }
 
 // ListSilenceds for use with mock lib
-func (c *MockClient) ListSilenceds(org string) ([]types.Silenced, error) {
-	args := c.Called(org)
+func (c *MockClient) ListSilenceds(org, sub, check string) ([]types.Silenced, error) {
+	args := c.Called(org, sub, check)
 	return args.Get(0).([]types.Silenced), args.Error(1)
 }