+++ types/event.go
@@ -1,6 +1,9 @@
 package types
 
-import "time"
+import (
+	"errors"
+	"time"
+)
 
 // EventFailingState indicates failing check result status
 const EventFailingState = "failing"
@@ -18,4 +21,17 @@ func FixtureEvent(entityID, checkID string) *Event {
 		Entity:    FixtureEntity(entityID),
 		Check:     FixtureCheck(checkID),
 	}
+}
+
+// Validate returns an error if the event does not pass validation tests.
+func (e *Event) Validate() error {
+	if err := e.Entity.Validate(); err != nil {
+		return errors.New("entity " + err.Error())
+	}
+
+	if err := e.Check.Validate(); err != nil {
+		return errors.New("check " + err.Error())
+	}
+
+	return nil
 }